better debug draw classes
